Added a proper AND and OR gate
Ok, the previous AND and OR gate implementaion was shit. Those were not
considered as proper logic gates because they will have a voltage drop
across them. So driving other similar gates with them will most probably
won't work.

Now the AND gate is an NAND gate with an inverter and OR gate is a NOR
gate with an inverter.